+++ src/constants.ts
@@ -16,8 +16,25 @@ export const FILTER_LIST_MARK = '#filters:';
 
 export const FILTER_EXT = '.txt';
 
-export const RECOMMENDED_TAG_ID = 1;
+export const RECOMMENDED_TAG_ID = 10;
 export const FILTER_LIST_URL = 'https://filters.adtidy.org/extension/chromium/filters.json';
+
+export type FilterType = {
+    filterId: number,
+    name: string,
+    description: string,
+    timeAdded: string,
+    homepage: string,
+    expires: number,
+    displayNumber: number,
+    groupId: number,
+    subscriptionUrl: string,
+    trustLevel: string,
+    version: string,
+    timeUpdated: string,
+    languages: string[],
+    tags: number[],
+};
 
 export type FilterListType = {
     groups: [
@@ -27,22 +44,5 @@ export type FilterListType = {
             displayNumber: number,
         },
     ],
-    filters: [
-        {
-            filterId: number,
-            name: string,
-            description: string,
-            timeAdded: string,
-            homepage: string,
-            expires: number,
-            displayNumber: number,
-            groupId: number,
-            subscriptionUrl: string,
-            trustLevel: string,
-            version: string,
-            timeUpdated: string,
-            languages: string[],
-            tags: number[],
-        },
-    ],
+    filters: FilterType[],
 };
