+++ src/fetchHelpers.ts
@@ -1,6 +1,7 @@
 import fetch, { Response } from 'node-fetch';
+import { FilterListType, FilterType, FILTER_LIST_URL } from './constants';
 
-export const fetchResponse = async (url: string): Promise<Response> => {
+const fetchResponse = async (url: string): Promise<Response> => {
     const response = await fetch(url);
     if (!response.ok) {
         const message = `Error status: ${response.status}. URL: ${url}`;
@@ -15,9 +16,22 @@ export const textFromResponse = async (url: string): Promise<string> => {
     return text;
 };
 
-// TODO fix any
-export const jsonFromResponse = async (url: string): Promise<any> => {
-    const response = await fetchResponse(url);
-    const json = await response.json();
-    return json;
+export const fetchFiltersByTag = async (tag: number) => {
+    const data = await textFromResponse(FILTER_LIST_URL);
+    const json: FilterListType = JSON.parse(data);
+    const recommended = json.filters
+        .filter((filter: FilterType) => filter.tags.includes(tag));
+
+    const urlList = recommended.map(
+        (filter: FilterType) => filter.subscriptionUrl,
+    );
+
+    const filters = await Promise.all(
+        urlList.map(async (url: string) => {
+            const text = await textFromResponse(url);
+            return text;
+        }),
+    );
+
+    return filters;
 };
