+++ src/index.ts
@@ -11,10 +11,9 @@ import {
     FILTER_EXT,
     RECOMMENDED_TAG_ID,
 } from './constants';
-import { fetchFiltersByTag } from './fetchFilters';
 import { github, imgur } from './api';
 import { getStringFromDescription, applyDiffToString } from './helpers';
-import { textFromResponse } from './fetchHelpers';
+import { textFromResponse, fetchFiltersByTag } from './fetchHelpers';
 import { screenshot } from './screenshot';
 import { extension } from './extension';
